tools/requirements.txt
certifi==2023.11.17
click==8.1.7
geojson==3.1.0
numpy==1.26.2
psycopg2-binary==2.9.9
pyproj==3.6.1
python-dotenv==1.0.0
shapely==2.0.2
